finish component rename
